Use GraphQL to search inventory
This allows us to get much richer records, with referenced items
dereferenced, and opens the way for us to request richer records still
when we're ready to transform them into OPAC records.

In order to make this work, we have to sanitise the Perl objects that
get deserialised from the JSON records to replace `JSON::PP::Boolean`
objects with regular 0/1, as `XMLout` chokes on the former.