Allow indexMap entries to be objects as well as strings.
When an object is provided, its `cql` member provides the mapping to
the CQL indexes, the same as the string value when that is provided.

The top-level `omitSortIndexModifiers` map is now abolished: instead,
each object-valued index-map entry can have its own
`omitSortIndexModifiers` member specifying the list of sort-modifiers
to omit.

(The purpose of this change is to make it possible to add further
items to the index-map specification.)