Tenant-specific configuration is now mandatory
If the named tenant has no configuration, then searches are rejected
with 109 Database unavailable.

This in accordance with the principle of least astonishment. While it
is mildly inconvenient to have to provide a vacuous tenant-specific
configuration file for each supported tenant, this avoids the much
worse problem of unexpected and hard-to-diagnose behaviour when trying
to use a tenant for which there is no specific configuration.