When checking for a barcode search recognise == as well as =
Addresses an issue raised in DEVOPS-558: when the `marcHoldings`
configuration entry `restrictToItem` is set to a true value, then only
the item nominated in the search has its details included. But the
recent change in the config, mapping `@attr 1=9998` to
`item.barcode==123" rather than "item.barcode=123", broke the check
for whether or not the search is for a single barcode. Now a query
that uses `=` or `==` -- or indeed `===` -- will work.