Clumsy approach to fixing post-processing
This includes doing the post-processing itself on the MARC::Record
object rather than the MARC-JSON structure, which means we can do
post-processing and holdings-and-item infos generation in either
order. But it still results in second and subsequent fetches doing the
holdings-and-items generation on a version of the MARC record that has
undergone post-processing -- hence a failure in test 07 when
generating such information from information in control field 008
which has been post-processed on a previous fetch.

I want to get this much neater before committing to master.