Finally, MARC holdings and post-processing work correctly together
This entailed major changes to internal APIs, and consequently a lot
of faffing about in the test scripts to make the tests run again.

At present, everything works (AND all tests pass, which is not the
same thing), but there are places where the code has a lot of
unnecessary local gnarliness. I'll fix that in the morning.