Replicate holdings-level location at the item level
From here, it can be inluded in MARC records by referencing the
`temporaryLocation` item-level field, which now falls back to this
value if the temporary location is not defined at that level; or by
referencing `_holdingsLocation` in any case.

The former is added as MARC subfield `952$l` in the standard and
sample Chicago configurations; but the latter is used in tests, so we
can see the resulting replication.

Fixes ZF-34.