Add `_permanentLocation` field to virtual item record
The virtual item records in holdings, from which OPAC records and
holdings-related MARC fields are generated, now contains not only
`temporaryLocation` but also the private field
`_permanentLocation`.

MARC holdings mappings can refer to this field, and the examples now
map it to subfield `L` of the 952 field.

An attempt to be all things to all people in ZF-43.