Field-substitution syntax changed to `%{245$a}`
We were previously using `${245$a}`, which is useless in the context
of a configuration file since the file-reader substitutes `${NAME}`
sequences with the values of environment variables before they ever
reach the point of being used in MARC post-processing.

Thanks to Wayne Schneider for finding this; and apologies for wasting
his time.

Properly fixes ZF-57.