Factor MARC-specificness out of transform functions
This was previous embedded in that we passed down both $marc (which we
are transforming from) and $newMarc (which we are building) through
all the transform functions, so that when we embed a field value in a
regular-expression replacement string using the `%{field}` syntax, we
can extract the necessary value.

Instead, we now pass down a closure that knows how to obtain the value
of a named field. For MARC transformations this works as before, but
for OPAC holdings transformations we can implement a different closure
with the same signature, allowing the actual transformation to work
identically for both formats.