Typo
Yes, I fixed a one-character typo in the change-log. But the real reason for this commit is so I can include in the git log the comment that I intended for the previous, much more substantial, commit -- but which got lost because I rushed through GitHub's merge process and didn't realise it had written its own ugly comment in place of the one I had written for the pull request. So here it is:

--

MARC postproc substitution from same field instance #39

When performing post-processing on MARC records, substitution replacement strings that contain sequences of the form `%{999$z}` are now treated slightly differently. If the field replacement value is to be drawn from another subfield of the field that is being modified, then the value is taken from the same instance of the field (as there may be several — as for example when the represent multiple items in holdings). When the value is to be drawn from a different field, then the first instance of that field is references as previously.

New test-cases added.

Fixes ZF-87.