Back down an unnecessarily recent dependency
When I added a dependency on HTTP::Cookies, I set it to 6.10 because
that's what was on the development machine I happened to be using. But
on another developement machine I only have 6.08, and that works just
fine. So I wound the dependency back to avoid an ugly (but benign)
warning.